+++ spec/policies/budget_policy_spec.rb
@@ -15,8 +15,7 @@ RSpec.describe BudgetPolicy do
       it { is_expected.to permit_action(:create) }
       it { is_expected.to permit_action(:edit) }
       it { is_expected.to permit_action(:update) }
-
-      it { is_expected.to forbid_action(:destroy) }
+      it { is_expected.to permit_action(:destroy) }
     end
 
     context "when the activity is a programme" do
@@ -26,8 +25,7 @@ RSpec.describe BudgetPolicy do
       it { is_expected.to permit_action(:create) }
       it { is_expected.to permit_action(:edit) }
       it { is_expected.to permit_action(:update) }
-
-      it { is_expected.to forbid_action(:destroy) }
+      it { is_expected.to permit_action(:destroy) }
     end
 
     context "when the activity is a project" do
@@ -151,8 +149,7 @@ RSpec.describe BudgetPolicy do
               it { is_expected.to permit_action(:create) }
               it { is_expected.to permit_action(:edit) }
               it { is_expected.to permit_action(:update) }
-
-              it { is_expected.to forbid_action(:destroy) }
+              it { is_expected.to permit_action(:destroy) }
             end
           end
         end